Refactoring
Code from index.js broken into multiple files and put in user_code folder. Added help function
